Added website communication properites to the Nodes and made a main.py to handel it. Some changes to NodeClass were also made.
Within the main.py I added a handler function , broadcast function, and main which looks for the IP address of the site (I used the example of 8080 but that should change to a more concrete port later on.)

I added node.web_listen to the Node class so that it can now listen for messages from the site by using asyncio. node_listen may need to be modified later on.